requirements.txt
# FastAPI e dependências principais
fastapi==0.114.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.1.0

# Cliente HTTP
httpx==0.27.0

# OpenAI
openai==1.51.2

# Processamento de dados
python-multipart==0.0.6

# Logging e monitoramento
structlog==23.2.0

# Validação e serialização
email-validator==2.1.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1

# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilitários
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
